refactor: Category 엔티티를 Course 엔티티에 병합
