+++ src/entities/course/model/category.type.ts
