+++ src/entities/course/model/course.type.ts
@@ -1,6 +1,6 @@
 import type { WriterType } from '@/src/entities/user'
 import type { CoursePlanPlaceType } from '@/src/entities/place/type'
-import type { CategoryType } from '@/src/entities/category/type'
+import type { CategoryType } from '@/src/entities/course/model'
 
 export type FavoriteRegionType = {
   id: number
