+++ src/entities/course/model/index.ts
@@ -1,1 +1,2 @@
-export * from './type'
+export * from './course.type'
+export * from './category.type'
